requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
blinker==1.6.2
captcha==0.5.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.0
defusedxml==0.8.0rc2
distro==1.9.0
Django==5.0.1
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-simple-captcha==0.6.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1
drf-yasg==1.21.7
Flask==2.3.2
frozenlist==1.4.0
greenlet==3.0.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
magic-filter==1.0.12
MarkupSafe==2.1.3
multidict==6.0.4
oauthlib==3.2.2
openai==1.10.0
openapi-codec==1.3.2
packaging==23.2
phonenumbers==8.13.27
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.0.3
pydantic_core==2.14.6
pygame==2.5.1
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
python-telegram-bot==20.3
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.16.2
simplejson==3.19.2
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
SQLAlchemy==2.0.19
sqlparse==0.4.4
stripe==8.4.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2
